phase/src/test/java/org/omnifaces/serve/renderkit/DefaultRenderKitFactoryTest.java
/*
 * Copyright (c) 2016 OmniFaces.org. All Rights Reserved.
 */
package org.omnifaces.serve.renderkit;

import javax.enterprise.inject.spi.CDI;
import org.jboss.weld.environment.se.Weld;
import org.jboss.weld.environment.se.WeldContainer;
import org.junit.After;
import static org.junit.Assert.assertNotNull;
import org.junit.Before;
import org.junit.Test;

/**
 * The JUnit tests for DefaultRenderKitFactory.
 */
public class DefaultRenderKitFactoryTest {

    /**
     * Stores the container.
     */
    WeldContainer container;

    /**
     * Stores the weld.
     */
    Weld weld;

    /**
     * Setup before testing.
     */
    @Before
    public void setUp() {
        weld = new Weld();
        container = weld.initialize();
    }

    /**
     * Tear down after testing.
     */
    @After
    public void tearDown() {
        weld.shutdown();
    }

    /**
     * Test getRenderKit method.
     */
    @Test
    public void testGetRenderKit() {
        RenderKitFactory factory = (RenderKitFactory) CDI.current().select(RenderKitFactory.class).get();
        assertNotNull(factory.getRenderKit("HTML_BASIC"));
    }
}
